requirements.txt
# ── Core RAG ──────────────────────────────────────────────────────────────────
rank-bm25>=0.2.2
numpy>=1.26,<2.0
python-dotenv>=1.0.0

# ── LLM / Embedding providers ─────────────────────────────────────────────────
openai>=1.25.0
httpx>=0.27.0
google-generativeai>=0.5.0
sentence-transformers>=3.0.0

# ── Config validation ─────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.2.0

# ── Retry / resilience ────────────────────────────────────────────────────────
tenacity>=8.2.3

# ── Token counting ────────────────────────────────────────────────────────────
tiktoken>=0.7.0

# ── Structured logging ────────────────────────────────────────────────────────
python-json-logger>=2.0.7

# ── Document loaders ──────────────────────────────────────────────────────────
pymupdf>=1.24.0
pdfplumber>=0.11.0
python-docx>=1.1.0
python-pptx>=0.6.23
pdf2image>=1.17.0
pytesseract>=0.3.10
beautifulsoup4>=4.12.0
requests>=2.32.0
pandas>=2.2.0
openpyxl>=3.1.2
unstructured>=0.14.0
pillow>=10.3.0

# ── Database — ORM + pgvector ─────────────────────────────────────────────────
sqlalchemy>=2.0.30
pgvector>=0.3.0
psycopg2-binary>=2.9.9

# ── Schema migrations ─────────────────────────────────────────────────────────
alembic>=1.13.0

# ── Reranking / local ML ──────────────────────────────────────────────────────
transformers>=4.41.0
torch>=2.3.0

# ── RAPTOR hierarchical indexing ──────────────────────────────────────────────
umap-learn>=0.5.6
scikit-learn>=1.5.0

# ── Agentic tools ─────────────────────────────────────────────────────────────
duckduckgo-search>=6.1.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0